backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# File Processing
ebooklib==0.18
python-magic==0.4.27

# Japanese Text Processing
mecab-python3==1.0.6
unidic-lite==1.0.8
jaconv==0.3.4

# AI/ML Integration
openai==1.3.5
requests==2.31.0

# Utilities
pydantic[email]==2.5.0
python-dotenv==1.0.0
aiofiles==23.2.1

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# CORS
fastapi-cors==0.0.6
